test: consolidate AHSP event-handling tests into a single class and BPMN
Merges `L4JAiAgentJobWorkerEventSubprocessTests` (regression-focused,
minimal BPMN) and the now-deleted `L4JAiAgentJobWorkerEventsTests`
(realistic, connectors BPMN) into a single `L4JAiAgentJobWorkerEventsTests`
running on the unified `agentic-ai-ahsp-connectors-event.bpmn`. Adds a
`Pending_Tool` service task as a deterministic in-AHSP park point, gives
the message subscription a variable-driven correlation key for per-test
isolation, and adds an output mapping to `SuperfluxProduct` (in both
`connectors.bpmn` and `connectors-event.bpmn`) so tool execution always
exercises the regression-sensitive code path.

Coverage: event before activation, event during execution
(WAIT_FOR_TOOL_CALL_RESULTS / INTERRUPT_TOOL_CALLS, with payload and
empty), and multi-event ordering. Each scenario asserts the chat
conversation, agent metrics and response text, the user-feedback worker
firing once, and that `toolCall` does not leak to the root scope —
camunda/camunda#51939. The defense-in-depth leak check in
`L4JAiAgentJobWorkerToolCallingTests` is now a real regression detector
thanks to the SuperfluxProduct output mapping.

All tests pass on the engine fix image and fail on 8.9.1.
